+++ src/main/java/cz/zsduhovacesta/service/database/StudentDao.java
@@ -130,7 +130,7 @@ public class StudentDao {
         student.setClassId(results.getInt(2));
         student.setLastName(results.getString(3));
         student.setFirstName(results.getString(4));
-        student.setFees(results.getDouble(5));
+        student.setFees(results.getInt(5));
         student.setVS(results.getInt(6));
         student.setMotherPhone(results.getString(7));
         student.setFatherPhone(results.getString(8));
@@ -139,9 +139,9 @@ public class StudentDao {
         student.setNotes(results.getString(11));
         student.setSchoolStage(results.getString(12));
         student.setPaymentNotes(results.getString(13));
-        student.setShouldPay(results.getDouble(14));
-        student.setPayed(results.getDouble(15));
-        student.setSummaryLastYear(results.getDouble(16));
+        student.setShouldPay(results.getInt(14));
+        student.setPayed(results.getInt(15));
+        student.setSummaryLastYear(results.getInt(16));
         return student;
     }
